Rename norman utils internal to norman core and norman utils external to norman utils.
(#86c778p0x)
