+++ norman_utils/codec_mapping.py
