+++ norman_utils/dot_syntax_resolver.py
