+++ norman_utils/encoding_combinations.py
