+++ norman_utils/encoding_defaults.py
