+++ norman_utils/encoding_utils.py
