+++ norman_utils/singleton.py
